[SAMZA-2787] GetDeleted API and Recover from DeletedException (#1676)
* GetDeleted API and Recover from DeletedException in commit for BlobStoreBackendFactory

* Fix style issues, fix breaking test cases

* Fixed test failure

* Fix failing integration test - move storeConsumer start after init() in ContainerStorageManager

* Style check fix

* Bug fixes, integration test

* Delete all blob from deleted SnapshotIndex

* clean up for final review

* Add integ test - delete snapshotindex and recover

* Fix failing integration tests related to init()

* Cleanup - remove unused code and move some code to util methods

* Unit test, minor refractoring

* Review comments - 1st round

* Refractor code
Remove taskcheckpoints mutation

* Address review comment

* Address review comment part 2

* Fix failing CSM test

* Addressed final review comments - more logs, consistent naming

* Checkstyle fix for test class

---------

+++ samza-test/src/test/java/org/apache/samza/test/util/TestBlobStoreManager.java
@@ -25,6 +25,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.CompletableFuture;
@@ -35,6 +36,7 @@ import org.apache.samza.checkpoint.CheckpointId;
 import org.apache.samza.config.Config;
 import org.apache.samza.storage.blobstore.BlobStoreManager;
 import org.apache.samza.storage.blobstore.Metadata;
+import org.apache.samza.storage.blobstore.exceptions.DeletedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,6 +49,7 @@ public class TestBlobStoreManager implements BlobStoreManager {
   public static final String LEDGER_FILES_DELETED = "filesRemoved";
   public static final String LEDGER_FILES_TTL_UPDATED = "filesTTLUpdated";
 
+  private final String deletedTombstone = "-DELETED";
   private final Path stateLocation;
   private final File filesAddedLedger;
   private final File filesReadLedger;
@@ -97,24 +100,45 @@ public class TestBlobStoreManager implements BlobStoreManager {
         metadata.getTaskName(), metadata.getStoreName(), suffix);
     LOG.info("Creating file at {}", destination);
     try {
-      FileUtils.writeStringToFile(filesAddedLedger, destination + "\n", Charset.defaultCharset(), true);
       FileUtils.copyInputStreamToFile(inputStream, destination.toFile());
+      FileUtils.writeStringToFile(filesAddedLedger, destination + "\n", Charset.defaultCharset(), true);
     } catch (IOException e) {
+      LOG.error("Error creating file: " + destination);
       throw new RuntimeException("Error creating file " + destination, e);
     }
     return CompletableFuture.completedFuture(destination.toString());
   }
 
   @Override
-  public CompletionStage<Void> get(String id, OutputStream outputStream, Metadata metadata) {
+  public CompletionStage<Void> get(String id, OutputStream outputStream, Metadata metadata, boolean getDeletedBlob) {
     LOG.info("Reading file at {}", id);
+    Path filePath = Paths.get(id);
     try {
-      FileUtils.writeStringToFile(filesReadLedger, id + "\n", Charset.defaultCharset(), true);
-      Path path = Paths.get(id);
-      Files.copy(path, outputStream);
+      Files.copy(filePath, outputStream);
       outputStream.flush();
+      FileUtils.writeStringToFile(filesReadLedger, id + "\n", Charset.defaultCharset(), true);
+    } catch (NoSuchFileException noSuchFileException) {
+      // Blob marked for deletion is suffixed with 'DELETED-'. Retrieve deleted blob if getDeletedBlob is True.
+      Path deletedFilePath = Paths.get(id + deletedTombstone);
+      String msg = String.format("File id: %s was not found. GetDeletedBlob was set to false", id);
+      if (!Files.exists(deletedFilePath)) {
+        throw new RuntimeException("404: " + msg, noSuchFileException);
+      }
+      if (getDeletedBlob) {
+        try {
+          Files.copy(deletedFilePath, outputStream);
+          outputStream.flush();
+          FileUtils.writeStringToFile(filesReadLedger, id + "\n", Charset.defaultCharset(), true);
+          LOG.info("Deleted File id {} found - Get deleted was set to true", id);
+        } catch (IOException e) {
+          throw new RuntimeException("Error reading file with GetDeleted set to true. File: " + id, e);
+        }
+      } else {
+        LOG.info("File id {} is deleted - Get deleted was set to false", id);
+        throw new DeletedException("410: " + msg, noSuchFileException);
+      }
     } catch (IOException e) {
-      throw new RuntimeException("Error reading file for id " + id, e);
+      throw new RuntimeException("Error reading file for id " + id + ". Get deleted was set to " + getDeletedBlob, e);
     }
     return CompletableFuture.completedFuture(null);
   }
@@ -124,7 +148,10 @@ public class TestBlobStoreManager implements BlobStoreManager {
     LOG.info("Deleting file at {}", id);
     try {
       FileUtils.writeStringToFile(filesDeletedLedger, id + "\n", Charset.defaultCharset(), true);
-      Files.delete(Paths.get(id));
+      // Suffix with 'DELETED' do not actually delete.
+      Files.move(Paths.get(id), Paths.get(id + deletedTombstone));
+    } catch (NoSuchFileException ex) {
+      LOG.warn("File might already be deleted. id: " + id);
     } catch (IOException e) {
       throw new RuntimeException("Error deleting file for id " + id, e);
     }
